Moved APIKeyAdminCreateForm from admin module to the forms module
